Feat: Validation and Tests

+++ app/Http/Controllers/TransactionsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Repositories\TransactionRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class TransactionsController extends Controller
 {
@@ -19,14 +20,48 @@ class TransactionsController extends Controller
         return $this->transactionsRepository->listExpense();
     }
     public function store(Request $request){
+        $rules = [
+            'amount' => 'required|numeric',
+            'description' => 'required|string',
+            'transaction_date' => 'required|date',
+            'type' => 'required|string|in:income,expense'
+        ];
+        $messages = [
+            'amount.required' => 'Amount is required',
+            'amount.numeric' => 'Amount must be a number',
+            'description.required' => 'Description is required',
+            'description.string' => 'Description must be a string',
+            'transaction_date.required' => 'Transaction date is required',
+            'transaction_date.date' => 'Transaction date must be a date',
+            'type.required' => 'Type is required',
+            'type.string' => 'Type must be a string',
+            'type.in' => 'Type must be income or expense'
+        ];
+        $validate = Validator::make($request->all(),$rules,$messages);
+        if($validate->fails()){
+            return response()->json([
+                'errors' => $validate->errors()
+            ],422);
+        }
+        
         return $this->transactionsRepository->store($request->all());
     }
     public function delete($id){
         $transaction = $this->transactionsRepository->find($id);
+        if($transaction == null){
+            return response()->json([
+                'message' => 'Transaction not found'
+            ],404);
+        }
         return $this->transactionsRepository->delete($transaction);
     }
     public function update(Request $request,$id){
         $transaction = $this->transactionsRepository->find($id);
+        if($transaction == null){
+            return response()->json([
+                'message' => 'Transaction not found'
+            ],404);
+        }
         return $this->transactionsRepository->update($transaction,$request->all());
     }
